+++ app/src/main/java/com/example/peterchu/watplanner/Database/DatabaseHandler.java
@@ -425,8 +425,10 @@ public class DatabaseHandler extends SQLiteOpenHelper {
         return ret;
     }
 
-    public CourseSchedule getCourseSchedule(String subject, String catalogNumber) {
+    public List<CourseComponent> getCourseSchedule(String subject, String catalogNumber) {
         SQLiteDatabase db = this.getReadableDatabase();
+
+        List<CourseComponent> ret = new ArrayList<>();
 
         Cursor cursor = db.query(
                 TABLE_SCHEDULES,
@@ -439,29 +441,15 @@ public class DatabaseHandler extends SQLiteOpenHelper {
                 null
         );
 
-        if (cursor == null) { return null; }
-        cursor.moveToFirst();
-        if (cursor.getCount() == 0) { return null; }
-
-        CourseSchedule courseSchedule = new CourseSchedule();
-        courseSchedule.setClassNumber(cursor.getInt(1));
-        courseSchedule.setSubject(cursor.getString(2));
-        courseSchedule.setCatalogNumber(cursor.getString(3));
-        courseSchedule.setTitle(cursor.getString(4));
-        courseSchedule.setEnrollmentCapacity(cursor.getInt(5));
-        courseSchedule.setEnrollmentTotal(cursor.getInt(6));
-        courseSchedule.setWaitingCapacity(cursor.getInt(7));
-        courseSchedule.setWaitingTotal(cursor.getInt(8));
-        courseSchedule.setType(cursor.getString(9));
-        courseSchedule.setSession(cursor.getString(10));
-        courseSchedule.setStartTime(cursor.getString(11));
-        courseSchedule.setEndTime(cursor.getString(12));
-        courseSchedule.setIsCancelled(cursor.getInt(13) == 1 ? true : false);
-        courseSchedule.setIsClosed(cursor.getInt(14) == 1 ? true : false);
-        courseSchedule.setIsTba(cursor.getInt(15) == 1 ? true : false);
-        courseSchedule.setDay(cursor.getString(16));
+        if (cursor.moveToFirst()) {
+            do {
+                CourseComponent courseComponent = makeCourseComponent(cursor);
+                ret.add(courseComponent);
+            } while (cursor.moveToNext());
+        }
         cursor.close();
-        return courseSchedule;
+
+        return ret;
     }
 
     public Course getCourseByCourseCode(String subject, String catalogNumber) {
